QAI-10345: Removed result int cast
With results upgraded to signed 64-bit int, the cast to int that was
originally used is no longer needed as sfnt_tsc_nsec returns int64_t.

Though, whilst the app should now be able to store the results properly,
it does not alleviate the issue of these large delays being present in
the first place.

+++ src/sfnt-pingpong.c
@@ -1069,7 +1069,8 @@ static void do_pings(int ss, int read_fd, int write_fd, int msg_size,
     sfnt_tsc(&start);
     do_ping(read_fd, write_fd, msg_size);
     sfnt_tsc(&stop);
-    results[i] = (int) sfnt_tsc_nsec(&tsc, stop - start - tsc.tsc_cost);
+    
+    results[i] = sfnt_tsc_nsec(&tsc, stop - start - tsc.tsc_cost);
     if( ! cfg_rtt )
       results[i] /= 2;
     if( cfg_sleep_gap )
